Different idea for sync: message contains index
This allows out-of-order sending, and parallelization of the Evaluator.
Previously all Evaluator threads would have to wait for the previous one
to finish before starting, since we didn't know how many messages would
be used in any index. Now, the postman can deliver incoming messages to
any waiting thread, regardless of its index.

Design wise, if the user wishes to not transmit the index (and save
1 byte of communication), they can handle sequencing the messages
themselves, and reconstruct the index on the evaluator's end.