significant change to API
Instead of just `send` and `receive`, we now split the OT trait into a sender
and a receiver. A new `init` function runs any one-time initialization, and then
`send` and `receive` could be used multiple times without having to keep
re-running any initialization. This is especially useful in OT extension.