Add a stopgap for a weird bug
Adding a println!() in the right place in the multiply function in the
evaluator was able to stop the test from breaking. For now, forcing a
function call to the hash appears to fix the problem. That being said, it
seems bizzare that the bug existed on both macOS and Linux, and is only caused
on nightly. While it's possible that this is caused by a compiler bug, I'm
hesitant to rule-out our use of unsafe code.