Fix and Re-enable GF(2^128)
I've manually implemented GF(2^128) multiplication. It's slightly slower
than before, since I've implemented the reduction using scalars instead
of vectors. If that becomes a problem we can change it. (The issue is
that there isn't a bitshift intrinsic for vectors, so it's not trivial
to rewrite the scalar code in terms of vectors.)

This commit also replaces quickcheck with proptest. Quickcheck's
implementation of Aribtrary for _all_ numeric types caps their maximum
value by the maximum size of a vector that it'd generate. As a result,
some tests were passing which should've been failing (because we were
expecting, for example, every bit of a random u128 to be random).