Add fused "lift into FiniteField and multiply"
This commit adds an operation to compute `lift(a)*b` where `a` is an
element of `FE::PrimeField` and `b` is an element of `FE`. We can
replace all instances of to_fpr() with this new operation.

This operation increases SVOLE performance on GF(2^128) by about 88%.

base_svole::extend::Gf128 (N = 1024)
    time:   [4.1712 ms 4.2075 ms 4.2530 ms]
    change: [-88.465% -88.321% -88.181%] (p = 0.00 < 0.05)
    Performance has improved.