Upgrade crates
The original packed_simd crate was breaking on nightly.

curve25519-dalek depends on rand 0.7. Rand 0.7 depends on packed_simd which is
broken (with the nightly feature). Rand 0.8 (with the nightly feature) depends
on a non-broken packed_simd. curve25519-dalek doesn't want to upgrade rand just
yet since it'd be a breaking change (as we are observing). So they want to wait
until there are more changes. So, what we can do is upgrade everything
else, stay on rand 0.7, and disable the nightly feature on rand 0.7.