Replace vectoreyes const macro with a fn
Previously, vectoreyes allowed users to declare a constant vector by
providing a macro like:

const THING: U32x8 = const_u32x4!([1,2,3,4]);

Internally, for the AVX2 backend, this macro expanded to transmute the
array of integers into the vector.

Unfortunately, (and I'm not quite sure why) this approach stopped
working.

Luckily, Rust recently stabilized using transmute from a const fn (the
reason why we'd initially used a macro was because, while transmute was
a stable const fn, it wasn't yet stable to call it from a const fn). As
a result, we can replace this macro with a const fn on vector types to
allow users of vectoreyes to construct constant vectors.