Add rustdoc script
In order to support LaTeX math in our docs, we take advantage of
rustdoc's --html-in-header flag. This flag is given the path of a file
to include in the header of the HTML. We have our etc/katex.html file
which contains the JavaScript neccessary to nicely display the math that
is in our rust docs.

What we'd like to do is set rustdoc's flags in .cargo/config to contain
the --html-in-header flag, pointing to our etc/katex.html file. The
problem is that we need to specify etc/katex.html as a path relative to
the workspace root (or at least relative to the .cargo/config file).
Unfortunately, we can only specify paths relative to the current
directory. If we specify etc/katex.html as the --html-in-header value in
the .cargo/config file, then cargo doc will work (when run from the root
directory), but cargo test will fail. This is because cargo test runs
rustdoc to run the doctests of the crate, and it does this from the
working directory of the crate that it's testing. As a result, the
rustdoc --html-in-header flag refers to a file that doesn't exist, and
rustdoc aborts.

To solve this, we add a etc/rustdoc.py file which sets up the flags
properly for generating latex-enabled docs, and we advise users to use
this instead of cargo doc (which will still work, it'll just be ugly).