Add random_nonzero method to finite fields
In a number of places in our codebase, we need to generate a random
non-zero field element. Rather than generically using rejection
sampling, some fields (such as F2) have more efficient implementations
of this operation. As a result, we add the random_nonzero method to the
FiniteField trait, to allow F2 (and other fields) to override this
behavior.