Add bit_decomposition method to finite fields
Previously, the VOLE code relied on the byte representation of a field
containing a bit decomposition of the field element, but that doesn't
need to be true in general.

This commit adds a bit_decomposition method which does exactly what VOLE
needs. It is also tested and has specific documentation, so that bit
decomposition will still be meaningful in, for example GF(3^135), where
there isn't an obvious interpretation of a bit.