Remove MODULUS from FiniteField
Some of the fields that we need to implement for the T&E team have
moduli which cannot fit in 128 bits. Rather than trying to pick a type
which can contain the modulus (and adding, for example, a dependency on
a big int library), we instead just store the number of bits needed to
represent the modulus. This is sufficient for most of our applications.