Remove WYKW specialization
This commit removes (most) of the code to specialize the WYKW VOLE to
different fields. I'm working on adding a new explicit round VOLE
implementation which can more easily specialized. Some specialization
code is kept intact as this new VOLE implementation leverages some of
the same functions (such as GGM) that the current implementation does.