Add decompose/from_subfield to finite fields
If A is a subfield of B, then it's always possible to decompose B into
a tuple of A values. And it's possible to turn a tuple of A values (of
the right length) into a B value. We'd previously only exposed this
functionality for prime subfields. This commit extends this
functionality for all subfields.