fb_reader.rs: Implement plugin type parsing in flatbuffers
This commit adds plugin type parsing to the flatbuffer reader. The
PluginTypeArgs type now has a constructor from &str, which
distinguishes between numeric and string arguments.

It would be nice to extract the common logic for reading decimal
integers into UInts, since it's duplicated in the text parser. Not
a huge deal for now.