lib.rs: Add visitor method for plugin function definitions
This commit adds define_plugin_function to the RelationVisitor trait,
and implements both the printing visitor and the flatbuffer parser
using it.

Note that PluginBindings are stored generically, and any decision-
making about what to do with them in more 'interesting' visitors (i.e.
that which is defined in reader.rs) is left to those implementors of
the visitor - this allows us to print plugin function bindings
easily/without concern for the particulars of the plugin under
consideration.