reader.rs: Handle calling mux functions
This commit adds support for calling mux functions to the function
body visitor for the sink. This emits MuxCall Instructions to be
interpreted in the plaintext evaluator and simple writer.

The MuxCall is guaranteed to have the input/output layout of a
spec-compliant mux function, if the call is valid (i.e. the wire
ranges match what's expected from the function definition.)