plaintext_eval.rs: Implement most of mux plaintext evaluation
This commit (mostly) implements the evaluation of mux calls in the
plaintext backend. Still TODO is converting the bits / single
field elements to a number that can be used to actually compute the
mux.

The trouble with this is we would like this number to be a Rust
usize, so it can be used to index into collections/process an
iterator. This is problematic, since in theory we may have a wire
carrying a value in a large field we support that needs 128 bits to
represent. In practice, it's unlikely people will be generating
such absurd muxes (I think), so this probably doesn't matter much,
but it's worth considering since the spec seems to allow for
arbitrarily large selectors.