simple_writer.rs: Implement binary condition logic
Implements the logic to compute the g vector for the binary case.
In particular, we compute g_i using a bit trick to check equality
of the condition wires and our counter i: We convert i to a Vec of
k bits, where k is the number of condition wires. The bits of i are
reversed (to get a big-endian representation) and flipped, at which
point we xor them bitwise with the condition wires. Finally, to
compute g_i itself, we and (multiply) these xor values.