reader.rs: Handle iteration plugin in define_plugin_function
This commit adds MapDefinitions to the global sink for the iteration
plugin methods, map and map_enumerated. These definitions will be
used in call() to, in the case of map, de-sugar into a series of
calls as determined by the iteration function definition.

In the case of map_enumerated, interpretation must be deferred to
the compiler backend (evaluator/writer), as wires must be allocated
for the counter - this cannot be safely done in the frontend.