reader.rs: Add wire-count checks to define_plugin_function.
As with mux, there are some conditions we can check up-front that
tell us whether an iteration function definition is valid or not.
This commit adds a couple of these checks (but not all, see below):
There must be exactly as many iterated-function outputs as there
are outputs to the function being iterated, and each corresponding
pair must have compatible counts based on the iter_count parameter.
That is, the outputs of the iterated function being defined must
have exactly iter_count times the number of wires in the
corresponding output of the closure.

A similar property is true of the non-environment inputs. For the
environment, the wire counts must match exactly.

It is also the case that all types must be consistent, but that is
not checked here, as it is expensive/requires structural equality
checking (since we are looking up the iterated closure and using
its saved type information, as used by the backends, rather than
the type index given by the relation's type definitions.) Any
problems resulting from type incompatibility will be triggered
down-stream (i.e. through calls to call() used to implement the
actual de-sugaring of iteration calls.)