reader.rs: De-sugar map calls
This commit implements the de-sugaring of map calls. It essentially
re-computes wire ranges based on the iteration count, and emits
FunctionCall instructions for these range windows.

This _should_ be refactored to reduce the duplication between this
case and the normal function call case. For the sake of testing,
however, this should be sufficient (and can be cleaned up before
merging.)