wire-map/src/lib.rs: Fix the implementation of wire deallocation.
Previously, the implementation of `free` on `WireMap` worked only
on wire ranges that covered a single allocation. The SIEVE IR0+
spec, however, allows for a range to be given to the `@delete`
directive that covers multiple allocations (as long as they are all
covered in full.)

This commit modifies the implementation of `free` to comply with
the spec. Largely the code is the same, but now wrapped in a loop
and featuring different checks on the allocation bounds.

Though FFMC and the compiler have a woeful lack of tests, I have
confirmed that this fixes the problems encountered with T&E's 3-SAT
circuits, and that all other T&E test cases pass as expected.