wire-map/src/lib.rs: Add allocation coverage check
Thanks to @benoit.razet for noting the need for this extra check.

The IR0+ spec includes:

> The range may span multiple allocations, but it must cover each
allocation in full; deallocating only part of an allocation is a
failure of resource validity.

I believe it's the case that the execution behavior of resource-
invalid circuits is undefined, and we have no obligation to report
such invalidities.

However, since we have a significant offline/compilation phase, it
is worth providing helpful error messages when resource (or other)
invalidities are encountered in the circuits we consume.

This particular specification of @delete poses some challenges - an
efficient allocation representation does not know where allocation
boundaries are in a range given to @delete (in particular, our
representation of allocated memory is keyed on the start of an
allocation, with the length of the allocation used as necessary to
determine endpoints.) This is why our implementation of @delete
simply walks the allocations, freeing everything that comprises the
given range.

Unfortunately, this means reporting this particular invalidity is
troublesome. On the bright side, we can only encounter this problem
at the very end of deallocation. If, for example, we have three
allocations $1...$10, $11...$20, and $21...$30, and we encounter
@delete($1...$29), we will deallocate the three aforementioned
ranges - but, this should be an error, since this is an invalid use
of delete. Prior to this commit, the deallocation would silently
just move on, deallocating more wires than asked for (and not
halting, despite the resource invalidity.)

This commit adds an extra check after the main deallocation loop.
For legitimate uses of @delete, our 'start' counter, curr_start,
ought to be exactly 1 more than the provided inclusive_end after
the loop terminates, since adding the final allocation length to
curr_start puts us right past the end of that allocation. This
check ensures that this part of the spec is properly checked,
and that we halt early on resource-invalid circuits, where the
source of the invalidity is a @delete that does not span complete
ranges.

This has been tested on a couple of the T&E circuits, and appears
to work as intended.