reader.rs: Fix check for wire range lengths in enumerated case.
This commit fixes a check implemented to support the map plugin
function. Specifically, previously, we did not account for the
single index wire range that is promised by the specification of
the iter plugin for map_enumerated. We now add an extra check for
the index wire range if we are reading a plugin function
definition.