circuit_ir.rs: Add an optional counter value to FunctionCall.
While the regular map operation can be fully unrolled/de-sugared in
the frontend of the compiler, the same is not true of map_enumerated.
Fully unrolling requires wire allocation, which is not something we
can safely emit at this stage, since we are streaming the circuit
and do not have access to the information we'd need to allocate on
wires that are not used in the circuit.

The strategy is still to emit the right number of FunctionCall,
which _does_ remove the iteration abstraction, but the mechanism
for passing the loop counter is pushed to the backend.

I believe that the question of plaintext evaluation is still
problematic, since that backend operates over the wire space of
SIEVE IR itself - it's unclear how to properly invoke the iterated
functions in that setting.

The same is not true for simple_writer - here, the counters will
manifest as outputs of Constant tasks, and there's no concern
about allocation in that case.

This is just a first attempt at map_enumerated - it's entirely
possible this solution is no good, and things will change :)