reader.rs: Add map_enumerated logic to call()
To account for map_enumerated, the implementation of the visitor
method call() had to be modified.

In particular, it was necessary to do two things:
1) For enumerated map, me must exclude the index wire range when
   calling make_ranges for the inputs. This wire range will be
   specially handled in the compiler backend, using...
2) The counter_value for enumerated maps must be Some(i), where i
   is the iteration counter. It should be an invariant of the
   compiler that _only_ this mapping case set counter_value to
   something other than None, as this is what flags to the backend
   that we must allocate task wire(s) for the counter.

All other logic remains the same - this unfortunately makes the
commit look more significant in terms of changes than it really is
since, previously, the logic for regular maps was enclosed in a
check that this was not an enumerated map, where really, this check
was needed in more particular places to modify the already-existing
logic. C'est la vie!