circuit_ir.rs: Add CounterInfo structure.
The compiler backend needs to know more than just the value of the
counter; it also needs to know in which field the counter wires
live, how many counter wires are being provided, and the number of
environment wire ranges that come before it. Since this was getting
complex, it made sense to wrap it all up in a single structure,
used as Option<CounterInfo> in FunctionCall.

Otherwise, this commit is uneventful, simply propagating the new
structure and computing its parts.