simple_writer.rs: Account for counter wires in function wire space.
To undo the erroneous calls emitted by map_enumerated, we must add
a 'gap' in the destination wires when calling borrow_child on the
wire map in the field of the counter wires at the point we have
gone past all environment wire ranges.

This is as simple as a conditional modification to input_pos before
assigning dst_start, since the wire range _actually_ at this index
should start num_wires later than if this was a regular function
call.

With this gap, we will now be able to compute the iteration counter
from its u64 value and allocate/assign to those wires.