circuit_ir.rs: num_env_for_field rather than num_env in CounterInfo.
A sublety missed in earlier commits was that, by the time we are
compiling a call, the WireRanges of inputs/outputs have been sorted
by field. This means that the gap we leave for counter wires must
be determined by the number of environment ranges _for a particular
field_.

This commit reflects this by replacing num_env in CounterInfo with
num_env_for_field. The calculation is in reader.rs (essentially a
filter of the first num_env elements of the inputs by type.)