simple_writer.rs: Implement counter wire allocation.
This commit implements the careful function wire allocation needed
for map_enumerated.

First, when computing the immutable ranges (i.e. inputs) in the
call to borrow_child, we conditionally advance the input position
based on whether this is a map_enumerated call or not. When it is,
we "make room" for num_wires by moving the input pointer ahead at
the right time (the field type matches and we have processed all
environment ranges for this field.)

After the child wiremap is borrowed, we conditionally allocate
space for the counter wires, and invoke the task-building
machinery needed to store the constants at the right locations.

In the case of binary, we must convert the counter to big-endian
bits, and store these in the wires. For arithmetic fields, we
know there will only be a single counter wire, so we store the
value directly in that one slot.

As of this writing, the compiler itself has been tested on a
small test circuit using map_enumerated, but this circuit has
not been run on the runner to confirm the proof succeeds. It
would also be a good idea to make sure no regressions were
introduced, particularly with 'normal' function calls (as this
plugin support touched that part of the code.)