Fix sccache in CI
It looks like sccache --start-server has been failing in CI due to a
timeout error. As a result, sccache has not been capturing the
environment variables and configuration of the sccache server, resulting
in sccache not using the shared build cache.

As a fallback, this commit additionally sets these same environment
variables at the CI level.