vector_v1.rs: 99% implementation of addc/mulc operations.
This commit (almost) implements all of addc and mulc for the vector
plugin.

The 'almost' relates to a limitation in DMC: We only support string
plugin arguments for now. This is problematic for the vector
plugin, as the operations involving constants specify those
constants in the plugin binding parameters (rather than requiring
a wire be allocated to pass to the function wrapper.)

A separate branch/MR will resolve this limitation in DMC.