plugins/mod.rs: Use PluginTypeArg instead of String for params.
Previously, DMC treated every plugin binding parameter as a string.
This is problematic for some plugins (permutation, vectors) that
deal with numeric parameters.

In particular, while it is probably sufficient in permutation to
assume that the tuple size fits in a u64 (I left a TODO in case we
want to revisit this assumption), for vectors, the operations that
add constant values require large integers we can represent as
bytes in the constant gates.

Additionally, the use of the text reader from the sieve-parser
crate would behave erroneously on circuits using permutation or
vectors, since the DMC implementation previously ignored all
numeric plugin arguments returned by that parser. This will be
discussed further and resolved in a later commit.