vector_v1.rs: Add test for each vector operation.
This commit adds a test for each of the vector operations.

The add test adds the vectors [-1, 1, -1] and [1, -1, 1].
The mul test multiplies the vectors [1, 0, 1] and [0, 1, 0].

The addc and mulc tests have a TODO dependent on MR !236, and fail.

The add_scalar test adds 1 to the vector [-1, -1, -1].
The mul_scalar test multiplies 0 by the vector [1, 1, 1].

The sum test computes the sums for each of the vectors [0],
[1, -1], [0, 1, -1], and [0, 1, -1. 0].

The product test computes the products for each of the vectors [0],
[0, 1], [0, 1, -1], and [0, 1, -1, 1].

The dotproduct test computes the dot products between:
[0]          and [1]
[0, 1]       and [1, 0]
[0, 1, 0]    and [1, 0, 1]
[0, 1, 0, 1] and [1, 0, 1, 0]

All tests except those for addc and mulc (see above) pass.