plugins/mod.rs: Remove the count parameter of gates_body.
This commit primarily removes the `count` parameter from
`gates_body`, and propogates this change to the mux and permutation
plugin implementations, being as minimally disruptive as possible
(i.e. not fully replacing the name 'count' in permutation.)

Additionally, the definition of `new_plugin` now uses the helper
function moved in the previous commit to compute the `args_count`
for the `FuncDecl`'s `CompiledInfo`.

The mux and permutation tests have been run; they all pass as
expected.