plugins/mod.rs: Pass FunStore to gates_body for plugin validation.
This commit adds a &FunStore parameter to the gates_body method in
order to support deeper validation of functions defined via the
iter plugin.

This is necessary due to the nature of the plugin - it refers to
other previously-defined functions in the circuit, and checking
that the signature of the iteration function is correct requires
access to the signature of the iterated function.

This commit also includes all upstream changes to new_plugin and
the text/FB parsers.