iter_v0.rs: Implement unrolling of iteration into calls.
This commit implements the details of unrolling the iteration into
calls.

The bulk of the work was computing the wire ranges to pass to the
calls; since we're unrolling at definition time rather than call
time, we must refer to inputs through the wire ID assignments
given by the spec's description of functions. This functionality
is provided as a helper function since the index math is a bit
weird.

There are a number of TODO items relating to the counter input used
in map_enumerated. In particular, we currently assume that the
counter input is a single wire always, though the spec allows for
any type/count of wires. The iter authors should be asked for
clarification about the intended use of the counter input; I would
have expected it to be like the condition wire from the mux plugin,
where only F2 allowed for multiple wires, and all other fields use
a single wire. If a wire can carry all of the finite field's values,
it does not make sense to me that for larger fields, there is any
value to multi-wire counter inputs (as big-endian values comprised
of these wires would necessarily exceed the maximum field value
when multiple wires are needed to represent.) The spec calls for
cycling back to 0, but this makes the multi-wire non-F2 case even
more odd to include.

This implementation does not yet have any tests - so it might be
completely wrong :) Proceed with caution!