iter_v0.rs: (Currently failing) tests for map and map_enumerated.
Add a test for each of map and map_enumerated. These tests are
taken from the spec/Python SIEVE-circuit-generating code.

Unfortunately the tests fail right now due to a panic while
evaluating the call into the iterator function - in particular,
there is an overflow when computing with the stack frames. I'm
currently trying to track down the source of this overflow; it's
definitely not an obvious problem with the circuit itself.