circuit_ir.rs: Include input/output types in type_presence.
Previously, the type_presence was computed by analyzing only the
gates of a plugin function. For iter, this is insufficient, since
in the case of map, we unroll to a collection of Call gates, which
do not provide any type information.

This commit adds Benoit's suggested fix: After analyzing the gates
for types, loop over the provided input and output counts, and set
those types as well.