field.rs: Add bigint conversions to the PrimeFiniteField trait.
This commit adds a constant MIN_LIMBS_NEEDED and two conversion
methods to the PrimeFiniteField trait.

The constant is the minimum number of word-sized values needed to
represent values in the `PrimeFiniteField`. I'm not sure if we can
make this portable - crypto_bigint chooses a type for limbs based
on the architecture's pointer width, so this constant will not be
the right value on 32-bit systems.

The two functions allow the caller to determine the width of the
integer being dealt with, and return `None` for conversions that
are impossible.

Finally, this commit adds stub implementations of the trait for
all PrimeFiniteFields.