field.rs: Compute MIN_LIMBS_NEEDED in the trait definition.
This commit uses the nlimbs! macro provided by crypto_bigint to
compute the constant MIN_LIMBS_NEEDED. This macro takes the number
of bits that need to be representable, which is 8 times the number
of bytes needed to represent the field value, which is given by the
CanonicalSerialize trait bound.

This calculation should be portable given the use of the macro.