field.rs: Add method to compute modulus as bigint.
This commit adds a method modulus_int to PrimeFiniteField that
returns the modulus of the field as a bigint with the given number
of limbs, if possible.

This commit also removes the constant definitions from the
implementations of the PrimeFiniteField trait, and adds stub
implementations of the modulus_int method.