field.rs: Update MIN_LIMBS_NEEDED documentation.
The MIN_LIMBS_NEEDED constant should indeed be large enough for
the field modulus; not just the field values.

The way we compute this value did not need to change; the
NumberOfBitsInBitDecomposition is already large enough for the
modulus, since for GF(p^r) it is defined as ceil(log_2(p))*r,
and ceil(log_2(p)) is the number of bits required to store p.

This commit reflects this in the documentation of
PrimeFiniteField and MIN_LIMBS_NEEDED.