f2.rs: Implement the PrimeFiniteField methods.
This commit replaces the stubs with actual implementations for the
field F2.

There isn't anything particularly tricky or notable about this
implementation; it is worth noting that there is no check in
try_from_int that the given integer does not exceed the value of
the field's modulus. This is fine, because we exhaustively check
the field elements (and the absolute minimum number of limbs
suffices to represent a single bit.)