field-test/src/field.rs: Add int conversion roundtrip test.
This commit adds a proptest to roundtrip crypto_bigint conversions.

Note that this commit will currently fail in CI due to some
odd implementation details for ff-generated prime fields; in fact,
all fields pass this roundtrip test besides F400p.

This fails due to an inconsistency between the number of bytes in
the modulus and the number of bytes specified in the macro
invocation that generates the field. The modulus fits in 50 bytes,
yet the macro specifies a num_bytes of 56.

My suggested fix is as follows:

1) The test_num_bytes property is not actually desirable; num_bytes
   should be the minimum number of bytes necessary to represent the
   modulus, regardless of how many internal limbs are required to
   represent the field elements, if the goal of canonical
   serialization is providing as-compact-as-possible byte
   representations of field elements.
2) The implementation of CanonicalSerialize should not use the
   actual_limbs value to compute slice bounds; rather, it should
   use Self::ByteReprLen (i.e. num_bytes.) This makes it clearer
   that all slices are of the correct size.

With these changes, all tests pass, including the crypto_bigint
roundtrip.