prime_field_using_ff.rs: Use ByteReprLen instead of limbs.
As noted in a previous commit, the implementations of
CanonicalSerialize for ff-generated fields were using the
$actual_limbs (multiplied by 8) for slice bounds in the to_ and
from_bytes methods. Furthermore, there was a test that asserted
ByteReprLen == $actual_limbs * 8.

If ByteReprLen is intended to be the number of bytes needed to hold
the modulus (specifically, the _minimum_ such number), this is not
correct.

This commit replaces the use of $actual_limbs in the aforementioned
slice bounds with ByteReprLen, and removes the erroneous test.
Furthermore, it corrects the macro invocation for F400p to use the
proper $num_bytes (i.e. 50).

With these changes, all tests on ff-generated fields once again
pass.