circuit_ir.rs: Introduce the DMCBigInt type (alias).
This commit adds a new type alias to the circuit IR module,
DMCBigInt. This type should alias Uint<LIMBS> with LIMBS large
enough for the largest field supported in DMC - for now, the
U384 (Uint<6>) type provided by crypto_bigint suffices.

If this type is used consistently where we need big integers,
changes to the type alias definition should not be breaking (unless
the change is to a too-small type.)