backend_multifield.rs: Bound with PrimeFiniteField.
The implementations of BackendT for the prover/verifier with
field conversion support previously bound FE with FiniteField,
with the additional constraint that the associated type
PrimeField be that same type FE.

This implies the bound FE: PrimeField, but unforutnately Rust
cannot determine this. This commit simply changes the bound on FE
to PrimeFiniteField, which would have been more correct in the
first place.