circuit_ir.rs: Use Number instead of Vec<u8>.
This is a very large commit that changes the gate representation
to operate over Numbers (cryptographic big integers) instead of
byte vectors.

We expect to see some performance gains from this change, since we
are removing the allocations associated with Vec (i.e. a layer of
indirection.)

This change required modifying many of the Diet Mac'n'Cheese
modulues, fortunately in usually-simple ways. Of note:

- The from_bytes_le method of BackendT is replaced with a method
  from_number, which is implemented for all backends in the
  obvious way (using the PrimeFiniteField methods.)
- The instance/witness values are also now handled as Numbers
  rather than bytes
- All tests still pass with this change