backend_trait.rs: Implement prime_field_value_from_number helper.
This commit removes some duplicated code in the implementations of
BackendT by adding a helper function prime_field_value_from_number
to the backend_trait module.

There is a somewhat open question about the FieldElement associated
type: Should it always be a PrimeFiniteField? If this is the case,
this helper function should be removed, the trait bound updated,
and the implementation moved into the definition of the trait as a
default. Extension fields seem to challenge this, however, and it's
possible we're already at the simplest solution that still allows
for that case.