backend_trait.rs: Add PrimeBackendT trait.
With the recent changes to Diet Mac'n'Cheese (particularly around
VOLE support), the previous work done to transition to the Number
type was disrupted, in particular because we've lost the bound
FieldElement: PrimeFiniteField from defining the field element as
the prime field of the top-level field.

To correct this (and, indeed, simplify the transition at the level
of these backends), this commit introduces the trait PrimeBackendT,
which may be defined on types that are BackendT.

The method from_number was moved from BackendT to this new trait.
Furthermore, a blanket implementation for all types T: BackendT
where T::FieldElement: PrimeFiniteField is provided - this actually
covers all of the the backend types defined in Diet Mac'n'Cheese,
leading to some duplicated code being removed.

Finally, to tie it all together, the BackendConvT trait is now
defined as a supertrait of PrimeBackendT rather than BackendT. This
is conceptually correct, as we indeed only implement the conversion
backends over PrimeFiniteFields.