Upgrade curve25519-dalek
Previously, we were using the curve25519-dalek-ng fork, because the
curve25519-dalek crate was stuck on an old version of the rand crate,
which was causing problems in Swanky. The latest release of the original
(unforked) library fixes these issues, so we've moved back to running
off this original library.