crates/party: Lift party crate out of mac-n-cheese.
The party crate (originally written to support ZK implementations)
defines abstractions useful to crates other than mac-n-cheese, and
contexts that are not ZK-specific.

This commit moves the party crate to the top-level shared crates/
alongside the field abstractions, flatbuffer-building code, and
other utilities used across all of swanky. All of the mac-n-cheese
crates have been updated to use this new top-level version of the
party crate.