Fix #147
jemalloc's configure script compiles code with -Werror, and Nix adds
-D_FORTIFY_SOURCE, which emits a warning if code is compiled without
optimization. This warning was upgraded to an error in the configure
script, which caused the configure script's detection mechnaisms to
fail.

To fix this problem, we force jemalloc to compile with optimization,
even on dev.

This shouldn't impact CI times since C builds are cached with sccache.