Refactor Swanky CI into external command
The old version of Swanky CI was not modular, and would only run in CI.
As we're adding more and more checks to the CI system, it would be nice
to (1) make it easier to add CI checks, and (2) be able to run CI checks
without needing to push to gitlab to do it.

This commit turns the old CI system into the ./swanky command, and
refactors the CI system into different modules for different lints.