cli.rs: Read some configuration from TOML file.
The number of command-line arguments to Diet Mac'n'Cheese has been
growing to the point of being unwieldy. In particular, we have
started exposing more "internal" options, such as the LPN size
for SVOLE and whether or not to use batching for assert-zero
checks.

This commit aims to keep the CLI manageable by moving some of these
options to a separate structure that is instead read from a TOML
configuration file.

The documentation for how to run DMC has been updated to reflect
this change (which is still largely untested).