cli.rs: Remove unnecessary derives, fix LPN configuration case.
The previous commit would fail to run with the configuration file
specified in the README, since the deserialization of the LpnSize
type is case-sensitive.

This commit fixes the documentation, and removes some unused code
from the cli module (namely, the ValueEnum derive macro, which is
no longer needed since the LPN size is not specified using clap
CLI arguments).