cli.rs: Remove the prover subcommand.
Previously, Diet Mac'n'Cheese used a CLI subcommand (ie something
like 'add' for git) to indicate prover-mode execution.

This was unnecessary, as the presence/absence of the --witness
argument is enough to determine whether to use prover or verifier
mode.

This commit removes that subcommand and updates the README to
reflect the change.