cli.rs: Open up the Config structure for more flexible config files.
Previously, the implementation required that the provided config
TOML file contained _all_ required keys. This is too inflexible.

This commit 'opens up' the configuration structure by making every
field optional. The Default implementation is updated to wrap each
of the values in Some, and a method to build a Config from a TOML
file is added. Since these are the only ways to construct a Config
value, we provide methods that unwrap the Option values since this
will always be safe.

The process of adding additional configuration parameters is given
in documentation comments for maintainers.