mac.rs: Use swanky-party and remove MacProver/MacVerifier.
This is a large commit, but largely consists of simple changes.
Most attention should be focused on mac.rs.

Diet Mac'n'Cheese has long featured the use of near-identical
structures separating prover and verifier data/functionality. This
has resulted in significant duplication of code, resulting in much
harder to maintain implementations.

This commit is the first step towards removing this duplication:
the swanky-party crate. In particular, this commit replaces the
types MacProver and MacVerifier with the single type Mac, which is
generic over the party.

Prover-specific capabilities are implemented as methods that
require evidence that the party is indeed the prover. Furthermore,
we were able to remove the Mac _trait_, since lift() can now be
implemented using the tools of the party crate - note the similarity
to the original implementation.

Making this change (and propagating it to correct all the type
errors) highlights where other parts of Diet Mac'n'Cheese can
benefit from use of the party abstraction. The goal of this MR will
be to remove as much duplicated code as possible, and this will
be the content of subsequent commits.