homcom.rs: Implement a party-generic FCom.
This is another big commit, but once again most of the interesting
stuff is in one module (homcom.rs), the changes to which are
detailed here.

This commit replaces the FComProver and FComVerifier types with
a single type, FCom, that is generic over the party.

We'll start with some type-level consequences: To enable proper
communication, FCom must be parametric over both the SVOLE sender
and receiver types, whereas these were local to the Prover/Verifier
structures previously. This does increase the verbosity of types
"up the chain", particularly for the field-conversion
evaluation structures. This may improve as we continue using
party-generic types.

In the code, there are some quirks: Where we previously didn't
need to provide, say, mutable output vectors, we now need them
regardless of the executing party. In the cases where the
old implementation would not write to this vector, the same is
true here - and such cases are clearly seen in the code through
the use of `&mut vec![]`.

The edabits module needed the most updating with these changes,
but they're largely straightforward.